+++ src/run/bddFixtures/index.ts
@@ -1,57 +1,22 @@
-import { TestInfo, test as base } from '@playwright/test';
-import { loadConfig as loadCucumberConfig } from '../cucumber/loadConfig';
-import { loadSteps } from '../cucumber/loadSteps';
-import { BddWorld, BddWorldFixtures, getWorldConstructor } from './bddWorld';
-import { BDDConfig, extractCucumberConfig } from '../config';
-import { getConfigFromEnv } from '../config/env';
-import { TestTypeCommon } from '../playwright/types';
-import { appendDecoratorSteps } from '../steps/decorators/steps';
-import { getPlaywrightConfigDir } from '../config/configDir';
-import { runScenarioHooks } from '../hooks/scenario';
-import { runWorkerHooks } from '../hooks/worker';
-import { IRunConfiguration } from '@cucumber/cucumber/api';
-import { StepInvoker } from './StepInvoker';
-import { ISupportCodeLibrary } from '../cucumber/types';
-import { TestMeta, TestMetaMap, getTestMeta } from '../gen/testMeta';
-import { logger } from '../utils/logger';
-import { getEnrichReporterData } from '../config/enrichReporterData';
-import { BddDataManager } from './bddData';
-import { loadStepsOwn } from '../cucumber/loadStepsOwn';
+import { test as base } from '@playwright/test';
+import { loadConfig as loadCucumberConfig } from '../../cucumber/loadConfig';
+import { loadSteps } from '../../cucumber/loadSteps';
+import { BddWorldFixtures, getWorldConstructor } from '../bddWorld';
+import { extractCucumberConfig } from '../../config';
+import { getConfigFromEnv } from '../../config/env';
+import { appendDecoratorSteps } from '../../steps/decorators/steps';
+import { getPlaywrightConfigDir } from '../../config/configDir';
+import { runScenarioHooks } from '../../hooks/scenario';
+import { runWorkerHooks } from '../../hooks/worker';
+import { StepInvoker } from '../StepInvoker';
+import { getTestMeta } from '../../gen/testMeta';
+import { logger } from '../../utils/logger';
+import { getEnrichReporterData } from '../../config/enrichReporterData';
+import { BddDataManager } from '../bddData';
+import { BddFixtures, BddFixturesWorker } from './types';
+import { loadStepsOwn } from '../../cucumber/loadStepsOwn';
 
 // BDD fixtures prefixed with '$' to avoid collision with user's fixtures.
-
-export type BddFixtures = {
-  // fixtures injected into BddWorld:
-  // empty object for pw-style, builtin fixtures for cucumber-style
-  $bddWorldFixtures: BddWorldFixtures;
-  $bddWorld: BddWorld;
-  Given: StepInvoker['invoke'];
-  When: StepInvoker['invoke'];
-  Then: StepInvoker['invoke'];
-  And: StepInvoker['invoke'];
-  But: StepInvoker['invoke'];
-  $testMetaMap: TestMetaMap;
-  $testMeta: TestMeta;
-  $tags: string[];
-  $test: TestTypeCommon;
-  $uri: string;
-  $scenarioHookFixtures: Record<string, unknown>;
-  $before: void;
-  $after: void;
-  $lang: string;
-};
-
-type BddFixturesWorker = {
-  $cucumber: {
-    runConfiguration: IRunConfiguration;
-    supportCodeLibrary: ISupportCodeLibrary;
-    World: typeof BddWorld;
-    config: BDDConfig;
-  };
-  $workerHookFixtures: Record<string, unknown>;
-  $beforeAll: void;
-  $afterAll: void;
-};
 
 export const test = base.extend<BddFixtures, BddFixturesWorker>({
   // load cucumber once per worker (auto-fixture)
@@ -84,7 +49,7 @@ export const test = base.extend<BddFixtures, BddFixturesWorker>({
   // init $bddWorldFixtures with empty object, will be owerwritten in test file for cucumber-style
   $bddWorldFixtures: ({}, use) => use({} as BddWorldFixtures),
   $bddWorld: async (
-    { $tags, $test, $bddWorldFixtures, $cucumber, $lang, $testMeta, $uri },
+    { $tags, $test, $step, $bddWorldFixtures, $cucumber, $lang, $testMeta, $uri },
     use,
     testInfo,
   ) => {
@@ -94,6 +59,7 @@ export const test = base.extend<BddFixtures, BddFixturesWorker>({
       supportCodeLibrary,
       $tags,
       $test,
+      $step,
       $bddWorldFixtures,
       lang: $lang,
       parameters: runConfiguration.runtime.worldParameters || {},
@@ -125,6 +91,8 @@ export const test = base.extend<BddFixtures, BddFixturesWorker>({
 
   // init $test with base test, but it will be overwritten in test file
   $test: ({}, use) => use(base),
+
+  $step: ({}, use) => use({ title: '' }),
 
   // feature file uri, relative to configDir, will be overwritten in test file
   $uri: ({}, use) => use(''),
@@ -175,14 +143,3 @@ export const test = base.extend<BddFixtures, BddFixturesWorker>({
     { auto: true, scope: 'worker' },
   ],
 });
-
-/** these fixtures automatically injected into every step call */
-export type BddAutoInjectFixtures = Pick<BddFixtures, '$test' | '$tags'> & {
-  $testInfo: TestInfo;
-};
-
-const BDD_AUTO_INJECT_FIXTURES: (keyof BddAutoInjectFixtures)[] = ['$testInfo', '$test', '$tags'];
-
-export function isBddAutoInjectFixture(name: string) {
-  return BDD_AUTO_INJECT_FIXTURES.includes(name as keyof BddAutoInjectFixtures);
-}